Add autoload switch
`parser` -> Autoload:
Default is `False`, can be changed in config to `True`.

`settings`:
The creation of the config has been redesigned, you need to call the
 `load` method, and accessing variables.

`fileSystem`:
Now the root will be generated in the tree, and not separately from
 the tree, that is, you will need to specify a relative path:
`Path('sets')` -> `Path('assets', 'sets')`
This helped avoid the following bug:
the config was created before the log.